alternative callback checking: shorter, more informative
For your consideration as a 'nicer' way to check for callbacks seen during
testing.  Especially for when the order of callbacks is known as it
additionally checks for that ordering, as well as the previous done "how many"
of each kind of callback.

Rather than a cumbersome counting of callbacks and then checking those counts,
a string accumulates each occurrence of each callback and then is checked at
the end of testing for the 'correct' match.

For example, if the order of callbacks should have been 'open_read_clos_unlk_'
but something else happened, the assert will trap with additional explanation:

  > [%   0|+  50|-   1]: fs_file_async
  > `fs_file_async` failed: exit code 134
  > Output from process `fs_file_async`:
  > callbacks seen were not as expected:
  >   expect 'open_read_clos_unlk_'
  >    found 'open_read_clos_buzz_'
  > Assertion failed in test/test-fs.c on line 445: cb_check("open_read_clos_unlk_")

This change touches only test/test-fs.c and only one test there, as a
demonstration of the idea.  Let me know what you think.